requirements.txt
aiofile==1.5.2
python-dotenv==0.10.3